Move and hide Cardano.Common.LocalSocket
Moved to Cardano.Node.LocalSocket and no longer an exposed module in
the library.
